# app/docs/swagger.yaml
---
swagger: "2.0"
info:
  version: "1.0.0"
  title: ""
host: "safnari.herokuapp.com"
basePath: "/api"
tags: []
schemes:
- "https"
paths:
  /users:
    post:
      parameters: 
      - in: body
        name: "body"
        schema:
          $ref: "#/definitions/user"
        required: true
      responses:
        201:
          description: User created
        400:
          description: Error
      tags:
        - users
    patch:
      parameters:
      - in: body
        name: "body"
        schema:
          $ref: '#/definitions/user'
      - in: header
        name: "Authorization"
        required: true
        type: string
      responses:
        200:
          description: User updated
        400:
          description: Error
      tags:
        - users
  /users/token:
    post:
      parameters:
        - in: body
          name: "body"
          required: true
          schema:
            type: object
            properties:
              email:
                type: string
              password:
                type: string
      tags:
        - users
      responses:
        200:
          description: "Successful"
        404:
          description: "Not found"
    delete:
      parameters:
        - in: header
          name: "Authorization"
          required: true
          type: string
      tags:
        - users
      responses:
        200:
          description: "Successful"
  /types:
    get:
      responses:
        200: 
          description: "Success"
      tags:
        - types
  /collections:
    get:
      parameters: 
        - in: header
          name: "Authorization"
          required: true
          type: string
        - in: query
          name: collection
          type: string
          description: "Collection id to filter by"
          required: false
      responses:
        200:
          description: "Success"
      tags:
        - collections
    post:
      parameters: 
        - in: body
          name: "body"
          schema:
            $ref: '#/definitions/collection'
      responses:
        200:
          description: "Success"
      tags:
        - collections
  /collections/{id}:
    get:
      parameters: 
        - in: header
          name: "Authorization"
          required: true
          type: string
        - in: path
          name: id
          type: string
          required: true
          description: "Collection id"
      responses:
        200:
          description: "Success"
      tags:
        - collections
    patch:
      parameters: 
        - in: body
          name: "body"
          schema:
            $ref: '#/definitions/collection'
        - in: path
          name: id
          type: string
          required: true
          description: "Collection id"
      responses:
        200:
          description: "Success"
      tags:
        - collections
  /items:
    post:
      parameters: 
        - in: header
          name: "Authorization"
          required: true
          type: string
        - in: body
          name: "body"
          schema:
            $ref: '#/definitions/item'
      responses:
        200:
          description: "Success"
      tags:
        - items
    get:
      parameters: 
        - in: header
          name: "Authorization"
          required: true
          type: string
      responses:
        200:
          description: "Success"
      tags:
        - items
  /items/{id}:
    get:
      parameters: 
        - in: header
          name: "Authorization"
          required: true
          type: string
        - in: path
          name: "id"
          required: true
          type: string
          description: "Item id"
      responses:
        200:
          description: "Success"
      tags: 
        - items
    patch:
      parameters: 
        - in: header
          name: "Authorization"
          required: true
          type: string
        - in: body
          name: "body"
          schema:
            $ref: '#/definitions/item'
        - in: path
          name: "id"
          required: true
          type: string
          description: "Item id"
      responses:
        200:
          description: "Success"
      tags:
        - items
    delete:
      parameters: 
        - in: header
          name: "Authorization"
          required: true
          type: string
        - in: path
          name: "id"
          required: true
          type: string
          description: "Item id"
      responses:
        200:
          description: "Success"
      tags: 
        - items
definitions:
  user:
    type: object
    properties:
      username:
        type: string
      picture:
        type: string
      email:
        type: string
      password:
        type: string
    required:
      - email
      - password
  type:
    type: object
    properties:
      name:
        type: string
      description:
        type: string
      icon:
        type: string
  collection:
    type: object
    properties:
      name: 
        type: string
        description: "Collection name"
      type:
        type: string
        description: "Type id"
      parent:
        type: string
        description: "Parent collection id"
  item:
    type: object
    properties:
      title:
        type: string
      number:
        type: number
      publisher:
        type: string
      artist:
        type: string
      format:
        type: string
      picture:
        type: string
      type:
        $ref: '#/definitions/type'
      coll:
        $ref: '#/definitions/collection'
      status:
        type: string
      
      
      
      
      
      
      
      
      
      
      
      
      
      
      
      
      
      
      
      
      
      
      
      
      
      
      